requirements.txt
python-binance
python-dotenv
flask
pytest